SA-358 Moved helper functions around
