+++ src/helpers/messageEvents.js
@@ -2,8 +2,7 @@
 import qs from 'query-string';
 import _ from 'lodash';
 import { getRelativeDates } from 'src/helpers/date';
-import { EVENTS_SEARCH_FILTERS} from 'src/constants';
-import { stringToArray } from 'src/helpers/string';
+import { EVENTS_SEARCH_FILTERS } from 'src/constants';
 
 /*
  * Translate the array of event definitions from /message-events/events/documentation
@@ -62,50 +61,6 @@ export function getDetailsPath(messageId, eventId){
 }
 
 /**
- * Creates the filters from the search query.
- *
- * @param {Object[]} searchQueries An array of query objects.
- * @returns {Object} An object containing the search parameters as keys and the search terms as values.
- */
-export function getFiltersFromSearchQueries(searchQueries = []) {
-
-  // Build a single object containing a key for each filter, initialised to an empty array
-  const emptyFilters = getEmptyFilters();
-
-  // Collect queries into an array of objects of form {key: value}
-  const queries = searchQueries.map(({key, value}) => ({[key]: stringToArray(value)}));
-
-  return Object.assign(emptyFilters, ...queries);
-}
-
-
-/**
- * Creates the search queries from the filter arrays.
- * @param {Object} filters An object containing the search parameters as keys and the search terms as values.
- * @returns {Object[]}  An array of query objects.
- */
-export function getSearchQueriesFromFilters(filters) {
-
-  //Filters out all search terms that are not part of the search query in AdvancedFilters
-  const { dateOptions, recipients, events, ...rest } = filters;
-
-  const nonEmptyFilters = removeEmptyFilters(rest);
-  const newSearchQueries = _.map(nonEmptyFilters, (value, key) => {
-    return ({ key, value: value.join(',')});
-    });
-  return newSearchQueries;
-}
-
-/**
- * Creates an array with all possible search filters as an object with a value=key and a label
- */
-export function getFiltersAsArray() {
-  const filtersAsArray = _.map(EVENTS_SEARCH_FILTERS, ({label}, key) =>
-    ({value: key, label}));
-  return filtersAsArray;
-}
-
-/**
  * Creates an object with all search filters set to an empty array
  */
 export function getEmptyFilters() {
@@ -114,8 +69,3 @@ export function getEmptyFilters() {
 
   return Object.assign({}, ...emptyFilters);
 }
-
-export function removeEmptyFilters(allFilters) {
-  return _.pickBy(allFilters, (value) => value.length>0);
-}
-
